script: Move debugger interfaces into `debugger/`
Part of #38901
