+++ components/script/dom/debugger/debuggerevalevent.rs
@@ -2,15 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use std::ptr::NonNull;
-
 use dom_struct::dom_struct;
-use js::jsapi::{Heap, JSObject, Value};
-use script_bindings::conversions::SafeToJSValConvertible;
-use script_bindings::reflector::DomObject;
 use script_bindings::str::DOMString;
 
-use crate::dom::bindings::codegen::Bindings::DebuggerAddDebuggeeEventBinding::DebuggerAddDebuggeeEventMethods;
+use crate::dom::bindings::codegen::Bindings::DebuggerEvalEventBinding::DebuggerEvalEventMethods;
 use crate::dom::bindings::codegen::Bindings::EventBinding::Event_Binding::EventMethods;
 use crate::dom::bindings::reflector::reflect_dom_object;
 use crate::dom::bindings::root::{Dom, DomRoot};
@@ -19,49 +14,42 @@ use crate::dom::types::{GlobalScope, PipelineId};
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
-/// Event for Rust → JS calls in [`crate::dom::DebuggerGlobalScope`].
-pub(crate) struct DebuggerAddDebuggeeEvent {
+/// Event for Rust → JS calls in [`crate::dom::debugger::DebuggerGlobalScope`].
+pub(crate) struct DebuggerEvalEvent {
     event: Event,
-    #[ignore_malloc_size_of = "Measured by the JS engine"]
-    global: Heap<*mut JSObject>,
+    code: DOMString,
     pipeline_id: Dom<PipelineId>,
     worker_id: Option<DOMString>,
+    frame_actor_id: Option<DOMString>,
 }
 
-impl DebuggerAddDebuggeeEvent {
+impl DebuggerEvalEvent {
     pub(crate) fn new(
         debugger_global: &GlobalScope,
-        global: &GlobalScope,
+        code: DOMString,
         pipeline_id: &PipelineId,
         worker_id: Option<DOMString>,
+        frame_actor_id: Option<DOMString>,
         can_gc: CanGc,
     ) -> DomRoot<Self> {
         let result = Box::new(Self {
             event: Event::new_inherited(),
-            global: Heap::default(),
+            code,
             pipeline_id: Dom::from_ref(pipeline_id),
             worker_id,
+            frame_actor_id,
         });
         let result = reflect_dom_object(result, debugger_global, can_gc);
-        result.event.init_event("addDebuggee".into(), false, false);
-
-        // Convert the debuggee global’s reflector to a Value, wrapping it from its originating realm (debuggee realm)
-        // into the active realm (debugger realm) so that it can be passed across compartments.
-        let cx = GlobalScope::get_cx();
-        rooted!(in(*cx) let mut wrapped_global: Value);
-        global
-            .reflector()
-            .safe_to_jsval(cx, wrapped_global.handle_mut(), can_gc);
-        result.global.set(wrapped_global.to_object());
+        result.event.init_event("eval".into(), false, false);
 
         result
     }
 }
 
-impl DebuggerAddDebuggeeEventMethods<crate::DomTypeHolder> for DebuggerAddDebuggeeEvent {
+impl DebuggerEvalEventMethods<crate::DomTypeHolder> for DebuggerEvalEvent {
     // check-tidy: no specs after this line
-    fn Global(&self, _cx: script_bindings::script_runtime::JSContext) -> NonNull<JSObject> {
-        NonNull::new(self.global.get()).unwrap()
+    fn Code(&self) -> DOMString {
+        self.code.clone()
     }
 
     fn PipelineId(
@@ -72,6 +60,10 @@ impl DebuggerAddDebuggeeEventMethods<crate::DomTypeHolder> for DebuggerAddDebugg
 
     fn GetWorkerId(&self) -> Option<DOMString> {
         self.worker_id.clone()
+    }
+
+    fn GetFrameActorId(&self) -> Option<DOMString> {
+        self.frame_actor_id.clone()
     }
 
     fn IsTrusted(&self) -> bool {
