+++ components/script/dom/debugger/debuggerframeevent.rs
@@ -15,7 +15,7 @@ use crate::dom::types::{GlobalScope, PipelineId};
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
-/// Event for Rust → JS calls in [`crate::dom::DebuggerGlobalScope`].
+/// Event for Rust → JS calls in [`crate::dom::debugger::DebuggerGlobalScope`].
 pub(crate) struct DebuggerFrameEvent {
     event: Event,
     pipeline_id: Dom<PipelineId>,
