+++ components/script/dom/debugger/mod.rs
@@ -1,0 +1,14 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod debuggeradddebuggeeevent;
+pub(crate) mod debuggerclearbreakpointevent;
+pub(crate) mod debuggerevalevent;
+pub(crate) mod debuggerframeevent;
+pub(crate) mod debuggergetenvironmentevent;
+pub(crate) mod debuggergetpossiblebreakpointsevent;
+pub(crate) mod debuggerglobalscope;
+pub(crate) mod debuggerinterruptevent;
+pub(crate) mod debuggerresumeevent;
+pub(crate) mod debuggersetbreakpointevent;
